separate layer_tests.py
